+++ examples/next15-app-router/shared/integrate-fedify.ts
@@ -12,11 +12,11 @@ import {
 import { keyPairsStore, relationStore } from "~/data/store";
 import { revalidatePath } from "next/cache";
 
-export const fedifyRequestHanlder = integrateFederation(() => {});
+export const fedifyRequestHandler = integrateFederation(() => {});
 
 const routePrefix = `/fedify-activity-handler`;
 
-const federation = createFederation<void>({
+const federation = createFederation({
   kv: new MemoryKvStore(),
 });