+++ CHANGES.md
@@ -33,12 +33,13 @@ To be released.
 ### @fedify/redis
 
  -  Added support for Redis Cluster to the *@fedify/redis* package.
-    [TODO]
+    [[#368] by Michael Barrett]
 
 [Next.js]: https://nextjs.org/
 [#313]: https://github.com/fedify-dev/fedify/issues/313
 [#353]: https://github.com/fedify-dev/fedify/issues/353
 [#365]: https://github.com/fedify-dev/fedify/pull/365
+[#368]: https://github.com/fedify-dev/fedify/pull/368
 
 
 Version 1.8.5